Split CamaraDeputadosHandler into two modules
